ai-llm-learning/lora_domain_learning_assistant/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
transformers>=4.40.0
peft>=0.11.0
trl>=0.8.6
accelerate>=0.30.0
datasets>=2.19.0
torch>=2.1.0
sentencepiece>=0.2.0
protobuf>=4.25.0
pytest>=8.0.0
# Optional QLoRA dependency. Enable config.model.use_4bit only when installed.
bitsandbytes>=0.43.0; platform_system != "Windows"
